Move generic packet writing code to MQTTSerializer
